+++ Source/DataSources/PolylineVolumeGraphics.js
@@ -24,7 +24,7 @@ import createPropertyDescriptor from './createPropertyDescriptor.js';
      * @param {Property} [options.outline=false] A boolean Property specifying whether the volume is outlined.
      * @param {Property} [options.outlineColor=Color.BLACK] A Property specifying the {@link Color} of the outline.
      * @param {Property} [options.outlineWidth=1.0] A numeric Property specifying the width of the outline.
-     * @param {Property} [options.shadows=ShadowMode.DISABLED] An enum Property specifying whether the volume casts or receives shadows from each light source.
+     * @param {Property} [options.shadows=ShadowMode.DISABLED] An enum Property specifying whether the volume casts or receives shadows from light sources.
      * @param {Property} [options.distanceDisplayCondition] A Property specifying at what distance from the camera that this volume will be displayed.
      *
      * @see Entity
@@ -154,7 +154,7 @@ import createPropertyDescriptor from './createPropertyDescriptor.js';
 
         /**
          * Get or sets the enum Property specifying whether the volume
-         * casts or receives shadows from each light source.
+         * casts or receives shadows from light sources.
          * @memberof PolylineVolumeGraphics.prototype
          * @type {Property}
          * @default ShadowMode.DISABLED