+++ Source/Scene/ShadowMode.js
@@ -1,7 +1,7 @@
 import freezeObject from '../Core/freezeObject.js';
 
     /**
-     * Specifies whether the object casts or receives shadows from each light source when
+     * Specifies whether the object casts or receives shadows from light sources when
      * shadows are enabled.
      *
      * @exports ShadowMode